Returns highest probability food in photo
