+++ GoRECIPE/gorecipe/lib/Screens/scan_screen.dart
@@ -2,12 +2,13 @@ import 'dart:async';
 import 'dart:io';
 // ignore: unused_import
 import 'dart:typed_data';
-
+import 'package:gorecipe/Screens/display_picture_screen.dart';
 import 'package:camera/camera.dart';
 import 'package:flutter/material.dart';
 import 'package:http/http.dart' as http;
 import 'package:http_parser/http_parser.dart';
-import 'dart:convert' show utf8;
+
+import 'dart:convert';
 
 // A screen that allows users to take a picture using a given camera.
 class ScanScreen extends StatefulWidget {
@@ -113,94 +114,6 @@ class ScanScreenState extends State<ScanScreen> {
           }
         },
         child: const Icon(Icons.camera_alt),
-      ),
-    );
-  }
-}
-
-// A widget that displays the picture taken by the user.
-class DisplayPictureScreen extends StatelessWidget {
-  final TextStyle style =
-      const TextStyle(fontFamily: 'Montserrat', fontSize: 20.0);
-  final String imagePath;
-
-  const DisplayPictureScreen({Key? key, required this.imagePath})
-      : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    upload(File img) async {
-      var postUri = Uri.parse(
-          "http://gorecipe.us-east-2.elasticbeanstalk.com/api/images/upload/4");
-      var request = http.MultipartRequest("POST", postUri);
-      request.files.add(http.MultipartFile.fromBytes(
-        'image',
-        await img.readAsBytes(),
-        filename: 'image.jpg',
-        contentType: MediaType('image', 'jpg'),
-      ));
-      // ignore: avoid_print
-      print("Request:");
-      // ignore: avoid_print
-      print(request);
-
-      var response = await request.send();
-      // ignore: avoid_print
-      print(response.statusCode);
-      response.stream.transform(utf8.decoder).listen((value) {
-        // ignore: avoid_print
-        print(value);
-      });
-    }
-
-    /*
-    getData() async {
-      print("In download");
-      var uri = Uri.http(
-          'www.gorecipe.us-east-2.elasticbeanstalk.com', '/api/users/2');
-
-      var request = http.MultipartRequest("GET", uri);
-
-      print("Request:");
-      print(request);
-      print(request.headers);
-
-      var response = await request.send();
-      print(response.statusCode);
-      response.stream.transform(utf8.decoder).listen((value) {
-        print(value);
-      });
-    }
-    */
-
-    final _nextbutton = Material(
-      elevation: 5.0,
-      borderRadius: BorderRadius.circular(30.0),
-      color: const Color.fromARGB(255, 116, 163, 126),
-      child: MaterialButton(
-        minWidth: MediaQuery.of(context).size.width / 4,
-        padding: const EdgeInsets.fromLTRB(10.0, 5.0, 10.0, 5.0),
-        onPressed: () {
-          upload(File(imagePath));
-          //getData();
-        },
-        child: Text("Next",
-            textAlign: TextAlign.center,
-            style: style.copyWith(
-                color: Colors.black, fontWeight: FontWeight.bold)),
-      ),
-    );
-
-    return Scaffold(
-      appBar: AppBar(title: const Text('Display the Picture')),
-      // The image is stored as a file on the device. Use the `Image.file`
-      // constructor with the given path to display the image.
-      body: Column(
-        children: [
-          Image.file(File(imagePath)),
-          const SizedBox(height: 40),
-          _nextbutton,
-        ],
       ),
     );
   }
