+++ GoRECIPE/gorecipe/pubspec.yaml
@@ -33,6 +33,7 @@ dependencies:
   camera: ^0.9.4
   http: ^0.13.4
   http_parser: ^4.0.0
+  shared_preferences: ^2.0.13
   path_provider:
   path:
